refactor: Article
