+++ src/main/java/toyboard/yoon/repository/ArticleRepository.java
@@ -1,10 +1,8 @@
 package toyboard.yoon.repository;
 
 import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Sort;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-import toyboard.yoon.domain.Article;
+import toyboard.yoon.entity.Article;
 
 import java.util.List;
